tokens y wallet

+++ database/migrations/2022_02_04_130746_create_tokens_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('symbol');
             $table->string('url',  2048)->nullable();
-            $table->integer('total_supply');
+            $table->bigInteger('total_supply'); 
             $table->foreignId('user_id')->references('id')->on('users');
             $table->timestamps();
         });
